ai-service/requirements.txt
crewai[tools]>=0.11.2
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
supabase>=2.13.0
python-dotenv>=1.1.0
pydantic>=2.10.0
